Rermoved local webpage and added remote site
