--- POM/homePagePOM.js
@@ -1,15 +1,0 @@
-import { Selector, t } from "testcafe";
-
-class homePage{
-
-    constructor(){
-        this.loginButton = Selector('#loginButton');
-    }
-
-    async clickLoginButton(){
-        await t
-            .click(this.loginButton);
-    }
-}
-
-export default new homePage();
