--- POM/modalPOM.js
@@ -1,15 +1,0 @@
-import { Selector, t } from "testcafe";
-
-class loginModal{
-
-    constructor(){
-        this.loginModal = Selector('#loginModal');
-    }
-
-    async clickLoginButton(){
-        await t
-            .click(this.loginButton);
-    }
-}
-
-export default new loginModal();
