+++ tests/login.js
@@ -1,8 +1,7 @@
 import { Selector, t, ClientFunction } from "testcafe"; 
-import homePage from "../POM/homePagePOM";
-import loginModal from "../POM/modalPOM";
+import homePage from "../POM/loginPOM";
 
-const url = '../index.html'
+const url = 'https://the-internet.herokuapp.com/login'
 
 fixture('Home Page')
     .page(url)
@@ -13,5 +12,5 @@ test('Click HomePage Login Button', async t => {
 
     homePage.clickLoginButton();
     await t
-        .expect(Selector(loginModal.loginModal).visible).ok();
+        .expect(Selector(homePage.invalidUsername).visible).ok();
 });
